-- consumption branch -- contributing into bd implemented

+++ app/src/main/java/com/freed_asd/fuel_calculator/data/local/consumption/city/ConsCity.kt
@@ -8,5 +8,5 @@ import androidx.room.PrimaryKey
 data class ConsCity (
     @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "consumption")val consumption: Float,
-    @ColumnInfo(name = "mileage")val mileage: Int
+    @ColumnInfo(name = "mileage")val mileage: Float
 )
